+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.2.0 (TBA)
 ### General
+* Fix [#58](https://g1cp.org/issues/58): Falling can no longer be interrupted by blocking mid-air.
 ### Story
 
 ## v1.1.0 (TBA)